Convert examples to a controlled component concept
